chore: moves config logic to its own file
To simplify upstreaming controller bits the whole Config struct and
related functions has been moved to its own file.

This will help upstreaming efforts as it keeps the logic isolated and
less vulnerable to merge conflicts moving forward.
